qcacmn: Check wmi service ext2 bit set with caution
Currently while checking for wmi service bit,  if
service bit is part of service ext2, host is not
validating if service bit index is valid or not,
and returns incorrect value.

For example, if host is checking for
WMI_SERVICE_PDEV_PARAM_IN_UTF_WMI and
num_wmi_service_ext_bitmap value is 4 then this service
bit can't be set, but seeing this value is set.

To address this issue if service ext2 bit
index is valid or not.

Change-Id: Ie4c766219e3793b46fa88dddb9544786142b92ef
CRs-Fixed: 3662190