qcacmn: Check and store both PSD and EIRP preferred support
The CTL transmit power data for each packet is based on the BDF for
each regulatory and is not based on the AFC response for SP power
modes. This may lead to transmit power violation or it may be
sub-optimal. So, for SP power mode, the transmit power should be
recomputed by HALPhy for all possible packet transmissions based on
the AFC response transmit power values sent by Host. If Host transmits
OOBE adjusted PSD and EIRP to HALPhy for all SP power mode or channel
changes, HALPhy will recompute CTL power for all possible packet
transmissions. in order to convey that HALPhy needs both PSD and EIRP,
a new service namely,
 WMI_SERVICE_BOTH_PSD_EIRP_FOR_AP_SP_CLIENT_SP_SUPPORT is advertised
 by the FW.

Read the service to check if both PSD and EIRP is preferred by the
target for SP and client SP power mode. Store it in regulatory psoc
objects. Write core and dispatcher APIs to store and read the
PSD and EIRP preferred support stored in regulatory psoc object.

Change-Id: Ic993c400dae4da314a588a55988673215cbcf407
CRs-Fixed: 3687366