qcacmn: Clear 40th bit of IPA DB address
It is common logic that the 40th bit of DB register need to be cleared
if WLAN chip is not attached to pcie slot 0.

Move related code to common area from within QCA_IPA_LL_TX_FLOW_CONTROL
which is only for WIN SP, and refine function name to look precise.

Change-Id: Iaaaa3341118189fcde78dba73c4de1e043d0de1e
CRs-Fixed: 3704102