qcacmn: Handle multiple filter deletion work in opt_dp_ctrl
IPA events are handled by a global workqueue,
and the work contexts which are posted to this
workqueue are also global, since it was expected
that two async events of the same type will never
come together. But with the addition of OPT_DP_CTRL,
two filter delete response can be received together
in an async manner and thereby can lead to the global
work context being over written.

To mitigate this issue, created two works for two cases i.e.
high throughput and filter delete requested.
Instead of scheduling work in loop we are scheduling work
one time with list of handle of filter deleted to
avoid overwrite.

Change-Id: Ibd35d4bb07ad4e456723aaeb8f4de41586e8e8e9
CRs-Fixed: 3750196