qcacmn: Clear number of users per ppdu in monitor mode
Currently, when a new ppdu is received, num_users per ppdu is
not getting cleared after every ppdu is processed. Hence, in OFDMA
cases the number of users increments for each ppdu and asserts when
number of users exceed the OFDMA_NUM_USERS.
To fix this, clear num_users when the ppdu is processed from status
ring in monitor mode.

Change-Id: If2598486338650342b83d65f2d5410a73c38c006
CRs-Fixed: 3664999