qcacmn: Update Reconfiguration ML IE definitions per 11be D7.0
Update definitions for Reconfiguration Multi-Link element as per
IEEE802.11be D7.0:
- Add definitions for more bits in the Presence Bitmap subfield in the
Reconfiguration Multi-Link element Control field.
- Clarify that the EML Capabilities subfield, the MLD Capabilities And
Operations subfield, and the Extended MLD Capabilities And Operations
subfield all in the Common Info field of the Reconfiguration
Multi-Link element have the same definitions as the corresponding
subfields in the Common Info field of the Basic Multi-Link element.
- Update max value in octets of Common Info Length subfield of Common
Info field in Reconfiguration variant Multi-Link element. Clarify that
this is the max value recognized. When parsing the Common Info Length
subfield, we may come across a value that is higher than the max
recognized value in case the sender implements a higher 11be draft,
or a higher specification. In such cases, we should ignore the bytes
in the Common Info field that we do not recognize.
- Update that a STA Profile may be present in the Per-STA Profile
subelement in the Reconfiguration Multi-Link element Link Info field.
- Update an existing definition and add definitions for subfields in
STA Control field of Per-STA Profile subelement in Reconfiguration
Multi-Link element Link Info field.
- Rename encoding for the 'Operation Update Type subfield' in STA
Control field of Per-STA Profile subelement in Reconfiguration
Multi-Link element Link Info field to encoding for the
'Reconfiguration Operation Type'. Add definitions for new encoding
values.
- Add definitions for encoding for NSTR Bitmap Size in STA Control
field of Per-STA Profile subelement in Reconfiguration Multi-Link
element Link Info field.
- Add definition for max size in octets of the NSTR Indication Bitmap
subfield in STA info field of Per-STA Profile subelement in
Reconfiguration Multi-Link element Link Info field.
- Update definition for max value in octets of STA Info Length in STA
Info field of Per-STA Profile subelement in Reconfiguration Multi-Link
element Link Info field. Clarify that this is the max value
recognized. When parsing the STA Info Length subfield, we may come
across a value that is higher than the max recognized value in case
the sender implements a higher 11be draft, or a higher specification.
In such cases, we should ignore the bytes in the STA Info field that
we do not recognize.

CRs-Fixed: 3918297
Change-Id: I87cae45f65b8dd2c89ab77d49d93eacf2061fa68