qcacmn: iounmap memory for the SOC which is going for a reset
During the umac_reset_interrupt_detach, the global ioremaped register
addresses are unmapped globally. In ERP test cases, one of the Radio
will be in the standby mode and since all the register values are cleared,
it is leading to the crash. Unmap the memory for the soc which is going for
a reset

Change-Id: I9a1db98e6205f4844d1c884e230ad33ea8c27ff8
CRs-Fixed: 3779517