qcacmn: handle filter delete from multiple context
This change is to handle filter delete from multiple
context in opt_dp_ctrl.

IPA events are handled by a global work queue,
and the work contexts which are posted to this
work queue are also global, since it was expected
that two async events of the same type will never
come together. But with the addition of OPT_DP_CTRL,
two filter delete response can be received together
in an async manner and thereby can lead to the global
work context being over written.

Earlier the implementation was to create two different works
for two cases high throughput and regular filter delete
to avoid overwritten work.

Changing implementation to use list for storing filter
handle deleted to be sent to WQ handler. Every handle
deleted will be queued to list and processed in handler.
This will avoid overwrite of same msg pointer from different
context.

Change-Id: I8da0ac59dbce89ee3c3a177021d92dbe71506435
CRs-Fixed: 3803587