qcacmn: Update power in 0.5 dBm for non-SP power types
The commit 'I43bc9abf4693e1222e5045b36f64c1904c85f834' added code
change for calculating the EIRP power for SP power mode in 0.5 dBm
resolution.

An additional argument was introduced to indicate whether the output
EIRP power will be in 0.5 dBm or 1 dBm. This argument is common for
all power modes. But the output EIRP power is scaled to 0.5 dBm only
for SP power mode. This caused the callers of this API to have
different behavior for different power modes.

To fix this, use 0.5 dBm resolution for non-SP power types too based
on the additional argument.

Change-Id: Ic4e28477ab6b22cbeaa6b315ba62d6563de842f5
CRs-Fixed: 3787068