qcacmn: Fix memory leak by freeing allocated msg memory
It was found that CCE filter release allocates memory
dynamically for msg to smmu unmap using work queue
and this memory has been freed in work queue handler but
it was not handled in else condition where there is no
need to unmap, causing memory leak.

This change is to fix this by freeing the memory
allocated in the case where msg is not handled in
work queue.

CRs-Fixed: 3814791
Change-Id: I3d1a011dcc164cb428fa5d8f272d22263fbcac10