qcacmn: Update chan width & score correctly for 320 MHz BW
Currently if link is 320 MHz capable chan width is set as 20
because handling of 320 MHz is not present while calculating
BW and default value 20 is assigned to chan width. While
calculating score of candidate with 320 MHz NSD is assigned
as 234 which is further used to find data rate. If a link is
having BW 160 MHz that link will score more because Nsd Value
will be 1980 for 160 MHz due to which NON-MLO candidate may
score more than MLO candidate and host may end up selecting
wrong candidate.

Fix is to add handling for 30 MHz while calculating BW and
while calculating ETP score.

CRs-Fixed: 3805876
Change-Id: I0fa7478a93d5e220bffb2f7a0edd5a39242fee9a