qcacmn: Fix invalid peer_id indicated in peer unmap notify to STC
Currently the peer_id from peer handle is passed to STC
when notified about peer unmap. This peer_id has been invalidated
before the notifier and hence the STC internal database never gets
cleared. If there is a subsequent peer mapped onto the same peer_id,
it leads to an assert.

To fix this issue, pass the cached peer_id to the peer unmap
notifier to STC.

Change-Id: I4199435a6e1240d0ff6763ed1c413c78d3859f63
CRs-Fixed: 3823907