qcacmn: Handle depletion in ipa free rx desc list
Depletion observed in rx desc free list used for
opt_dp_ctrl in absence of regular RX buffer cause
free Rx desc from regular buffer is used to replenish
this list of descriptors.

To avoid this depletion in opt_dp_ctrl rx desc list,
FW is sending a dummy buffer on every ctrl pkt sent and
descriptor used from this list with special marking in peer
metadata of the buffers.

This change is to handle that dummy buffer and replenish
rx desc to the list for every ctrl pkt.

Change-Id: Ieeaad528b86e99010210fdb2ea74f4fac14ec348
CRs-Fixed: 3837426