qca-wifi-oss: EIR: Repeater is not connecting back to Root after Root is up
After Root is rebooted or disconnected, RE1 is trying to
connect to RE2 and updated the preferred bssid as well.
Since RE2 is already connected with RE1, it rejects the
connection and never connects to Root once it is back.

Since RE2 didnot update the extender_info(that it lost
the Root connection), the above issue is seen.

The extender_info update failed because of wrong handling
of mac_addr and bssid comparison check. Fixed the condition.

Change-Id: I17f3aad767d0dcf05af94b7f760cc2ab2e5dfef3