qcacmn: Add score based on EHT mode of operation
Previously, the Host driver assigned a minimum score of 1 to
any candidate, if the BSSID was in the deny list.

This commit introduces a change to prioritize candidates based on
their link type. The new scoring system adds more weight to MLO
over SLO and Legacy links. The priority order is as follows:
MLO 3-link > MLO 2-link > SLO > Legacy.

Change-Id: I1bb8247d7a2ae88967c0949c0a51e32a3d8a44da
CRs-Fixed: 3855155