qcacmn: Fix MODE0 recovery issue while WSI remap
During WSI remap make sure the mode0 recovery is supported
A chip that is in the bypass state should remain in the bypass
post recovery also and user should be able to bring it back
out of bypass. The recovery dumps will be triggered for the
asserted socs. Since the bypassed SOC will be in down state
that will not be asserted

Change-Id: I1aad42a935a1e6c872dc58528dcf83ffaf50d639
CRs-Fixed: 3832366