qca-wifi: Service class delete when MSDUQ is reclaimed
Currently a Service class cannot be deleted once
a MSDUQ is instantiated and associated with it.

As we have MSDUQ reclaim feature now where once
the flow stops, the used queue can be reclaimed,
the service class can also be deleted when we
receive a SCS delete from client after no flows
are using it. Also, the service class delete
case is handled when a peer gets deleted also.

SAWF delete cases are also handled.

Change-Id: I556a95095fcafa18cb8a344079f41f8907d93003