qcacmn: Fix transmitted BSSID missing for nontransmitted partner link
The issue arises when dealing with nontransmitted BSSID during beacon
scanning. the raw frame of a nontransmitted BSSID scan entry inherits
information from the transmit BSSID beacon, but the multiple BSSID IE is
removed. Consequently, when an MLO associated link is connected, a new
scan entry is created from the raw frame of the partner link to avoid scan
entry aging out, But the MBSSID information is lost. The transmitted
BSSID isn't sent to the firmware when the nontransmitted partner link vdev
up, leading to beacon miss.

To address this issue, after the MLO associated link is connected, we
should renew the timestamp of the partner link scan entry to prevent scan
cache aging out. However, we should refrain from updating other
information, such as the MBSSID info.

Change-Id: I14cd16e7a9243eb2dfa8ea540c056d994f248413
CRs-Fixed: 3846461